+++ RedPandaIDE/settings.cpp
@@ -2640,6 +2640,10 @@ static void setDebugOptions(Settings::PCompilerSet pSet) {
 }
 
 bool Settings::CompilerSets::addSets(const QString &folder, const QString& cc_prog) {
+    foreach (const PCompilerSet& set, mList) {
+        if (set->binDirs().contains(folder))
+            return false;
+    }
     // Default, release profile
     PCompilerSet baseSet = addSet(folder,cc_prog);
     if (!baseSet)